restored 1 command per nudge
