+++ manage_remote.py
@@ -190,7 +190,7 @@ def get_users_with_no_hours(env):
     if env.upper() == "PROD":
         click.secho("this is running things on the (live) server!", fg="red")
 
-    task = run(env, "email_users",  "--cmd", "users_with_no_hours")
+    task = run(env, "users_with_no_hours")
     click.echo(task)
 
 
@@ -200,7 +200,7 @@ def get_users_with_seven_hours_no_survey(env):
     if env.upper() == "PROD":
         click.secho("this is running things on the (live) server!", fg="red")
 
-    task = run(env, "email_users", "--cmd", "users_with_seven_hours_no_survey")
+    task = run(env, "users_with_seven_hours_no_survey")
     click.echo(task)
